Move `gatsby-ts` and `utils` into `src` directory
