+++ gatsby-config.js
@@ -1,4 +1,4 @@
 require('source-map-support').install();
 require('ts-node').register();
 
-module.exports = require('./gatsby-ts/gatsby-config');
+module.exports = require('./src/gatsby/gatsby-config');
