+++ src/gatsby/gatsby-browser/onRouteUpdate.ts
