+++ src/gatsby/gatsby-node/pages/createApiSpecPages/utils/getPrimitive.spec.ts
