+++ src/gatsby/gatsby-node/pages/createApiSpecPages/utils/normalizeArray.spec.ts
