+++ src/gatsby/gatsby-node/pages/createApiSpecPages/utils/normalizeArray.ts
