+++ src/gatsby/gatsby-node/pages/createApiSpecPages/utils/normalizeObject.spec.ts
