+++ src/gatsby/gatsby-ssr/index.ts
