+++ src/types/remark-slug/index.d.ts
